SIGTERM handler
We already handle the SIGINT, but when we receive a SIGTERM, we just
exit the main Avocado process and leave the test subprocess behind.

This patch creates a SIGTERM proper handler, to kill the test process
and set the INTERRUPTED status to the job.

Reference: https://trello.com/c/6qBdaSM8
Signed-off-by: Amador Pahim <apahim@redhat.com>